+++ crates/api/src/site/purge/post.rs
@@ -11,7 +11,7 @@ use lemmy_api_common::{
 use lemmy_db_schema::{
   source::{
     local_user::LocalUser,
-    moderator::{AdminPurgePost, AdminPurgePostForm},
+    mod_log::admin::{AdminPurgePost, AdminPurgePostForm},
     post::Post,
   },
   traits::Crud,
