+++ crates/apub/src/activities/block/undo_block_user.rs
@@ -27,7 +27,7 @@ use lemmy_db_schema::{
   source::{
     activity::ActivitySendTargets,
     community::{CommunityPersonBan, CommunityPersonBanForm},
-    moderator::{ModBan, ModBanForm, ModBanFromCommunity, ModBanFromCommunityForm},
+    mod_log::moderator::{ModBan, ModBanForm, ModBanFromCommunity, ModBanFromCommunityForm},
     person::{Person, PersonUpdateForm},
   },
   traits::{Bannable, Crud},
