+++ crates/db_schema/src/source/mod_log/mod.rs
@@ -1,0 +1,2 @@
+pub mod admin;
+pub mod moderator;
