+++ crates/db_schema/src/source/mod_log/moderator.rs
@@ -1,10 +1,6 @@
 use crate::newtypes::{CommentId, CommunityId, PersonId, PostId};
 #[cfg(feature = "full")]
 use crate::schema::{
-  admin_purge_comment,
-  admin_purge_community,
-  admin_purge_person,
-  admin_purge_post,
   mod_add,
   mod_add_community,
   mod_ban,
@@ -299,96 +295,4 @@ pub struct ModAddForm {
   pub mod_person_id: PersonId,
   pub other_person_id: PersonId,
   pub removed: Option<bool>,
-}
-
-#[skip_serializing_none]
-#[derive(Clone, PartialEq, Eq, Debug, Serialize, Deserialize)]
-#[cfg_attr(feature = "full", derive(Queryable, Selectable, Identifiable, TS))]
-#[cfg_attr(feature = "full", diesel(table_name = admin_purge_person))]
-#[cfg_attr(feature = "full", diesel(check_for_backend(diesel::pg::Pg)))]
-#[cfg_attr(feature = "full", ts(export))]
-/// When an admin purges a person.
-pub struct AdminPurgePerson {
-  pub id: i32,
-  pub admin_person_id: PersonId,
-  #[cfg_attr(feature = "full", ts(optional))]
-  pub reason: Option<String>,
-  pub when_: DateTime<Utc>,
-}
-
-#[cfg_attr(feature = "full", derive(Insertable, AsChangeset))]
-#[cfg_attr(feature = "full", diesel(table_name = admin_purge_person))]
-pub struct AdminPurgePersonForm {
-  pub admin_person_id: PersonId,
-  pub reason: Option<String>,
-}
-
-#[skip_serializing_none]
-#[derive(Clone, PartialEq, Eq, Debug, Serialize, Deserialize)]
-#[cfg_attr(feature = "full", derive(Queryable, Selectable, Identifiable, TS))]
-#[cfg_attr(feature = "full", diesel(table_name = admin_purge_community))]
-#[cfg_attr(feature = "full", diesel(check_for_backend(diesel::pg::Pg)))]
-#[cfg_attr(feature = "full", ts(export))]
-/// When an admin purges a community.
-pub struct AdminPurgeCommunity {
-  pub id: i32,
-  pub admin_person_id: PersonId,
-  #[cfg_attr(feature = "full", ts(optional))]
-  pub reason: Option<String>,
-  pub when_: DateTime<Utc>,
-}
-
-#[cfg_attr(feature = "full", derive(Insertable, AsChangeset))]
-#[cfg_attr(feature = "full", diesel(table_name = admin_purge_community))]
-pub struct AdminPurgeCommunityForm {
-  pub admin_person_id: PersonId,
-  pub reason: Option<String>,
-}
-
-#[skip_serializing_none]
-#[derive(Clone, PartialEq, Eq, Debug, Serialize, Deserialize)]
-#[cfg_attr(feature = "full", derive(Queryable, Selectable, Identifiable, TS))]
-#[cfg_attr(feature = "full", diesel(table_name = admin_purge_post))]
-#[cfg_attr(feature = "full", diesel(check_for_backend(diesel::pg::Pg)))]
-#[cfg_attr(feature = "full", ts(export))]
-/// When an admin purges a post.
-pub struct AdminPurgePost {
-  pub id: i32,
-  pub admin_person_id: PersonId,
-  pub community_id: CommunityId,
-  #[cfg_attr(feature = "full", ts(optional))]
-  pub reason: Option<String>,
-  pub when_: DateTime<Utc>,
-}
-
-#[cfg_attr(feature = "full", derive(Insertable, AsChangeset))]
-#[cfg_attr(feature = "full", diesel(table_name = admin_purge_post))]
-pub struct AdminPurgePostForm {
-  pub admin_person_id: PersonId,
-  pub community_id: CommunityId,
-  pub reason: Option<String>,
-}
-
-#[skip_serializing_none]
-#[derive(Clone, PartialEq, Eq, Debug, Serialize, Deserialize)]
-#[cfg_attr(feature = "full", derive(Queryable, Selectable, Identifiable, TS))]
-#[cfg_attr(feature = "full", diesel(table_name = admin_purge_comment))]
-#[cfg_attr(feature = "full", diesel(check_for_backend(diesel::pg::Pg)))]
-#[cfg_attr(feature = "full", ts(export))]
-/// When an admin purges a comment.
-pub struct AdminPurgeComment {
-  pub id: i32,
-  pub admin_person_id: PersonId,
-  pub post_id: PostId,
-  #[cfg_attr(feature = "full", ts(optional))]
-  pub reason: Option<String>,
-  pub when_: DateTime<Utc>,
-}
-
-#[cfg_attr(feature = "full", derive(Insertable, AsChangeset))]
-#[cfg_attr(feature = "full", diesel(table_name = admin_purge_comment))]
-pub struct AdminPurgeCommentForm {
-  pub admin_person_id: PersonId,
-  pub post_id: PostId,
-  pub reason: Option<String>,
 }
