+++ crates/db_views_moderator/src/structs.rs
@@ -8,22 +8,21 @@ use lemmy_db_schema::{
     federation_allowlist::AdminAllowInstance,
     federation_blocklist::AdminBlockInstance,
     instance::Instance,
-    moderator::{
-      AdminPurgeComment,
-      AdminPurgeCommunity,
-      AdminPurgePerson,
-      AdminPurgePost,
-      ModAdd,
-      ModAddCommunity,
-      ModBan,
-      ModBanFromCommunity,
-      ModFeaturePost,
-      ModHideCommunity,
-      ModLockPost,
-      ModRemoveComment,
-      ModRemoveCommunity,
-      ModRemovePost,
-      ModTransferCommunity,
+    mod_log::{
+      admin::{AdminPurgeComment, AdminPurgeCommunity, AdminPurgePerson, AdminPurgePost},
+      moderator::{
+        ModAdd,
+        ModAddCommunity,
+        ModBan,
+        ModBanFromCommunity,
+        ModFeaturePost,
+        ModHideCommunity,
+        ModLockPost,
+        ModRemoveComment,
+        ModRemoveCommunity,
+        ModRemovePost,
+        ModTransferCommunity,
+      },
     },
     person::Person,
     post::Post,
